# swagger_server/models/reception_data.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server.models.reception_detail_data import ReceptionDetailData  # noqa: F401,E501
from swagger_server import util


class Receptiondata(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, dispatch_id: int=None, is_correct: bool=None, reception_details: List[ReceptionDetailData]=None, user: str=None, observations: str=None):  # noqa: E501
        """Receptiondata - a model defined in Swagger

        :param dispatch_id: The dispatch_id of this Receptiondata.  # noqa: E501
        :type dispatch_id: int
        :param is_correct: The is_correct of this Receptiondata.  # noqa: E501
        :type is_correct: bool
        :param reception_details: The reception_details of this Receptiondata.  # noqa: E501
        :type reception_details: List[ReceptionDetailData]
        :param user: The user of this Receptiondata.  # noqa: E501
        :type user: str
        :param observations: The observations of this Receptiondata.  # noqa: E501
        :type observations: str
        """
        self.swagger_types = {
            'dispatch_id': int,
            'is_correct': bool,
            'reception_details': List[ReceptionDetailData],
            'user': str,
            'observations': str
        }

        self.attribute_map = {
            'dispatch_id': 'dispatch_id',
            'is_correct': 'is_correct',
            'reception_details': 'reception_details',
            'user': 'user',
            'observations': 'observations'
        }
        self._dispatch_id = dispatch_id
        self._is_correct = is_correct
        self._reception_details = reception_details
        self._user = user
        self._observations = observations

    @classmethod
    def from_dict(cls, dikt) -> 'Receptiondata':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Receptiondata of this Receptiondata.  # noqa: E501
        :rtype: Receptiondata
        """
        return util.deserialize_model(dikt, cls)

    @property
    def dispatch_id(self) -> int:
        """Gets the dispatch_id of this Receptiondata.


        :return: The dispatch_id of this Receptiondata.
        :rtype: int
        """
        return self._dispatch_id

    @dispatch_id.setter
    def dispatch_id(self, dispatch_id: int):
        """Sets the dispatch_id of this Receptiondata.


        :param dispatch_id: The dispatch_id of this Receptiondata.
        :type dispatch_id: int
        """

        self._dispatch_id = dispatch_id

    @property
    def is_correct(self) -> bool:
        """Gets the is_correct of this Receptiondata.


        :return: The is_correct of this Receptiondata.
        :rtype: bool
        """
        return self._is_correct

    @is_correct.setter
    def is_correct(self, is_correct: bool):
        """Sets the is_correct of this Receptiondata.


        :param is_correct: The is_correct of this Receptiondata.
        :type is_correct: bool
        """

        self._is_correct = is_correct

    @property
    def reception_details(self) -> List[ReceptionDetailData]:
        """Gets the reception_details of this Receptiondata.


        :return: The reception_details of this Receptiondata.
        :rtype: List[ReceptionDetailData]
        """
        return self._reception_details

    @reception_details.setter
    def reception_details(self, reception_details: List[ReceptionDetailData]):
        """Sets the reception_details of this Receptiondata.


        :param reception_details: The reception_details of this Receptiondata.
        :type reception_details: List[ReceptionDetailData]
        """

        self._reception_details = reception_details

    @property
    def user(self) -> str:
        """Gets the user of this Receptiondata.


        :return: The user of this Receptiondata.
        :rtype: str
        """
        return self._user

    @user.setter
    def user(self, user: str):
        """Sets the user of this Receptiondata.


        :param user: The user of this Receptiondata.
        :type user: str
        """

        self._user = user

    @property
    def observations(self) -> str:
        """Gets the observations of this Receptiondata.


        :return: The observations of this Receptiondata.
        :rtype: str
        """
        return self._observations

    @observations.setter
    def observations(self, observations: str):
        """Sets the observations of this Receptiondata.


        :param observations: The observations of this Receptiondata.
        :type observations: str
        """

        self._observations = observations