# swagger_server/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: zent-dispatch-ms
  description: Dispatch microservice
  version: 1.0.0
servers:
- url: /rest/zent-dispatch-api/v1.0/
paths:
  /status-dispatch:
    get:
      tags:
      - Catalogs
      summary: Obtiene todos los estados de despacho
      description: Devuelve todos los estados de la base
      operationId: get_status_dispatch
      parameters:
      - name: externalTransactionId
        in: header
        required: true
        style: simple
        explode: false
        schema:
          type: string
          example: fcea920f7412b5da7be0cf42b8c93759
      - name: channel
        in: header
        required: true
        style: simple
        explode: false
        schema:
          type: string
          example: ZENTINEL
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
        "503":
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
      x-openapi-router-controller: swagger_server.controllers.dispatch_controller.DispatchView
  /dispatch-products-sku:
    get:
      tags:
      - Catalogs
      summary: Obtiene todos los productos de despacho
      description: Devuelve todos los productos de despacho de la base
      operationId: get_dispatch_products
      parameters:
      - name: externalTransactionId
        in: header
        required: true
        style: simple
        explode: false
        schema:
          type: string
          example: fcea920f7412b5da7be0cf42b8c93759
      - name: channel
        in: header
        required: true
        style: simple
        explode: false
        schema:
          type: string
          example: ZENTINEL
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
        "503":
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
      x-openapi-router-controller: swagger_server.controllers.dispatch_controller.DispatchView
  /vehicle-types:
    get:
      tags:
      - Catalogs
      summary: Obtiene todos los tipos de vehículo
      description: Devuelve todos los tipos de vehículo de la base
      operationId: get_vehicle_types
      parameters:
      - name: externalTransactionId
        in: header
        required: true
        style: simple
        explode: false
        schema:
          type: string
          example: fcea920f7412b5da7be0cf42b8c93759
      - name: channel
        in: header
        required: true
        style: simple
        explode: false
        schema:
          type: string
          example: ZENTINEL
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
        "503":
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
      x-openapi-router-controller: swagger_server.controllers.dispatch_controller.DispatchView
  /reception:
    post:
      tags:
      - Dispatch
      summary: Guarda la recepcion del despacho en la base de datos.
      description: Guardado de recepcion de despacho
      operationId: post_reception
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestReception"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
        "503":
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
      x-openapi-router-controller: swagger_server.controllers.dispatch_controller.DispatchView
  /dispatch:
    get:
      tags:
      - Dispatch
      summary: Obtiene todos los despachos
      description: Devuelve todos los despachos de la base
      operationId: get_dispatch
      parameters:
      - name: user
        in: header
        required: false
        schema:
          type: string
          example: "dmales"
      - name: start_date
        in: query
        required: false
        schema:
          type: string
          format: date-time
          example: "2025-01-21T00:00:00"
      - name: end_date
        in: query
        required: false
        schema:
          type: string
          format: date-time
          example: "2025-01-22T00:00:00"
      - name: externalTransactionId
        in: header
        required: true
        schema:
          type: string
          example: "fcea920f7412b5da7be0cf42b8c93759"
      - name: channel
        in: header
        required: true
        schema:
          type: string
          example: "ZENTINEL"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
        "503":
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
      x-openapi-router-controller: swagger_server.controllers.dispatch_controller.DispatchView
    post:
      tags:
      - Dispatch
      summary: Guarda el despacho en la base de datos.
      description: Guardado de despacho
      operationId: post_dispatch
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestDispatch"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
        "503":
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
      x-openapi-router-controller: swagger_server.controllers.dispatch_controller.DispatchView
  /dispatch/{dispatch_id}:
    patch:
      tags:
      - Dispatch
      summary: Actualiza el despacho en la base de datos.
      description: Actualiza de despacho
      operationId: update_dispatch
      parameters:
      - name: dispatch_id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          example: 1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestDispatch"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
        "503":
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
      x-openapi-router-controller: swagger_server.controllers.dispatch_controller.DispatchView
components:
  schemas:
    GenericResponse:
      properties:
        data:
          type: object
          nullable: true
          example: Estructura generica.
        errorCode:
          type: integer
          example: 0
        message:
          type: string
          example: Mensaje generico.
        internalTransactionId:
          type: string
          example: d290f1ee-6c54-4b01-90e6-d701748f0851
        externalTransactionId:
          type: string
          example: fcea920f7412b5da7be0cf42b8c93759
      example:
        errorCode: 0
        internalTransactionId: d290f1ee-6c54-4b01-90e6-d701748f0851
        externalTransactionId: fcea920f7412b5da7be0cf42b8c93759
        message: Mensaje generico.
        data: {}
    ResponseError:
      type: object
      properties:
        errorCode:
          type: integer
        internalTransactionId:
          type: string
        externalTransactionId:
          type: string
        message:
          type: string
    RequestGeneric:
      properties:
        channel:
          type: string
          example: ZENTINEL
        externalTransactionId:
          type: string
          example: fcea920f7412b5da7be0cf42b8c9375
    RequestDispatch:
      properties:
        dispatch_data:
          $ref: "#/components/schemas/RequestDispatch_dispatch_data"
        channel:
          type: string
          example: ZENTINEL
        externalTransactionId:
          type: string
          example: fcea920f7412b5da7be0cf42b8c93759
    RequestDispatch_dispatch_data_products_sku:
      type: object
      properties:
        id_product:
          type: integer
          example: 1
        quantity:
          type: number
          example: 10
    RequestDispatch_dispatch_data:
      type: object
      properties:
        weight:
          type: number
          format: double
          example: 2000000.75
          nullable: true
        vehicle_type:
          nullable: true
          type: integer
          example: 1
        sku_type:
          nullable: true
          type: string
          example: Multiple
        destiny:
          nullable: true
          type: integer
          example: 1
        status_id:
          type: integer
          nullable: true
          example: 1
        truck_license:
          nullable: true
          type: string
          example: GME-511
        user:
          nullable: true
          type: string
          example: dmales
        driver:
          nullable: true
          type: string
          example: Daniel
        images:
          nullable: true
          maxItems: 10
          type: array
          items:
            type: string
            format: binary
        products_sku:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/RequestDispatch_dispatch_data_products_sku"
        observations:
          type: string
          nullable: true
          example: N/A
    RequestReception:
      properties:
        reception_data:
          $ref: "#/components/schemas/Receptiondata"
        channel:
          type: string
          example: ZENTINEL
        externalTransactionId:
          type: string
          example: fcea920f7412b5da7be0cf42b8c93759
    Receptiondata:
      type: object
      properties:
        dispatch_id:
          type: integer
          example: 1
        is_correct:
          type: boolean
          example: true
        reception_details:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/ReceptionDetailData"
        user:
          type: string
          example: dmales
        observations:
          type: string
          nullable: true
          example: N/A
    ReceptionDetailData:
      type: object
      properties:
        reception_id:
          type: integer
          nullable: true
          example: 1
        product_id:
          type: integer
          example: 1
        expected_quantity:
          type: integer
          example: 1
        received_quantity:
          type: integer
          example: 1
        observations:
          type: string
          nullable: true
          example: N/A